UX audit: fix WhatsApp link, add skip-to-content, improve pricing language, mobile schedule, and SEO meta tags
- Fix broken WhatsApp link on contact page (was missing Dara's number)
- Add skip-to-content accessibility link in site layout
- Show price ranges on collapsed offering cards
- Replace "income-based model" with "pay what feels right" (Charles Eisenstein style gift economy)
- Update contribution tier descriptions to remove income-range language
- Improve community schedule mobile readability (stack date/time on small screens)
- Add per-page metadata for offerings, community, the-rose, guardians, contact
- Improve global SEO keywords (rose meditation, aura reading course, etc.)
- Refactor pages into server/client split for Next.js metadata support
