+++ src/app/(site)/guardians/page.tsx
@@ -1,62 +1,25 @@
-'use client';
+import type { Metadata } from 'next';
+import GuardiansClient from './GuardiansClient';
 
-import { useRef } from 'react';
-import { motion, useInView } from 'framer-motion';
-import { guardians, brandQuotes } from '@/lib/data';
-
-import PageHero from '@/components/sections/PageHero';
-import GuardianCard from '@/components/sections/GuardianCard';
-import QuoteBlock from '@/components/sections/QuoteBlock';
-import InvitationCTA from '@/components/sections/InvitationCTA';
-
-// =============================================================================
-// GUARDIANS PAGE
-// =============================================================================
+export const metadata: Metadata = {
+  title: 'Guardians',
+  description:
+    'Meet the guardians of ROSES OS — Angelina Ataíde, Dara Ayoub, Diego Dosal, and Peggy Mars. Teachers, stewards, and practitioners carrying the Rose Meditation and Aura Reading lineage.',
+  keywords: [
+    'rose meditation teachers',
+    'aura reading teachers',
+    'ROSES OS guardians',
+    'spiritual teachers',
+    'energy healing mentors',
+  ],
+  openGraph: {
+    title: 'Guardians | ROSES OS',
+    description:
+      'Meet the guardians who carry and transmit the Rose Meditation and Aura Reading lineage.',
+    url: '/guardians',
+  },
+};
 
 export default function GuardiansPage() {
-  const gridRef = useRef<HTMLElement>(null);
-  const gridInView = useInView(gridRef, { once: true, margin: '-100px' });
-
-  return (
-    <>
-      {/* 1. Hero */}
-      <PageHero
-        eyebrow="Guardians"
-        title="The Keepers of the Rose Field"
-        description="The Guardians are devoted practitioners and stewards of the Rose technology. Each one brings a unique gift to the field, holding space, guiding practice, and nurturing the community with presence and care."
-        image="/page-images/page-guardians.png"
-      />
-
-      {/* 2. Guardian Cards Grid */}
-      <section ref={gridRef} className="section-padding">
-        <div className="container-premium">
-          <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-8 lg:gap-12">
-            {guardians.map((guardian, i) => (
-              <motion.div
-                key={guardian.id}
-                initial={{ opacity: 0, y: 24 }}
-                animate={gridInView ? { opacity: 1, y: 0 } : {}}
-                transition={{
-                  duration: 0.6,
-                  delay: i * 0.12,
-                  ease: [0.16, 1, 0.3, 1],
-                }}
-              >
-                <GuardianCard guardian={guardian} />
-              </motion.div>
-            ))}
-          </div>
-        </div>
-      </section>
-
-      {/* 3. Quote */}
-      <QuoteBlock
-        quote={brandQuotes[3].text}
-        variant="fullbleed"
-      />
-
-      {/* 4. Invitation CTA */}
-      <InvitationCTA />
-    </>
-  );
+  return <GuardiansClient />;
 }
