+++ src/lib/seo.tsx
@@ -7,21 +7,27 @@ import { Metadata } from 'next';
 
 export const siteConfig = {
   name: 'ROSES OS',
-  description: 'A seamless path to inner freedom. ROSES OS is a living ecosystem of practices, teachings, and community designed to help you remember what you already know and live from that place.',
+  description: 'A seamless path to inner freedom. ROSES OS offers Rose Meditation and Aura Reading courses online — a living ecosystem of practices, teachings, and community to help you remember who you truly are.',
   url: process.env.NEXT_PUBLIC_SITE_URL || 'https://rosesos.com',
   ogImage: '/og-image.jpg',
   creator: '@rosesos',
   keywords: [
+    'rose meditation',
+    'aura reading course',
+    'aura reading online',
+    'energy healing',
     'consciousness',
-    'remembrance',
-    'coherence',
     'inner freedom',
-    'meditation',
-    'sacred technology',
-    'the rose',
-    'human development',
     'spiritual practice',
+    'meditation course online',
+    'chakra cleansing',
+    'aura cleansing',
+    'clairvoyance training',
+    'coherence',
+    'remembrance',
+    'sacred technology',
     'holistic wellness',
+    'spiritual development',
   ],
 };
